Created New Sorted Data
I took our csv files, and made new ones that I think may be a bit more useful as I added additional statistics that will be useful for us (i.e. previous cumulative runs scored/allowed, average runs scored/allowed). With the way the previous data worked, each game appeared twice in the dataframe once for each team. Thus, if we were to use this we would be simulating games twice, so I cut out half of them, and relabeled the columns in terms of a team1 and team2, so no data was removed, but the number of rows is now halved.